+++ src/Compute/Compute/ChangeLog.md
@@ -21,6 +21,8 @@
 -->
 ## Upcoming Release
 * Edited Get-AzVm to filter by `-Name` prior to checking for throttling due to too many resources. 
+* Edited `New-AzSnapshot` cmdlet to check for existing snapshot with the same name in the same resource group. 
+    - Throws an error if a duplicate snapshot exists. 
 
 ## Version 4.6.0
 * Added `-VmssId` parameter to `New-AzVm`